Improve tests and fixtures for testing templates
Previously, `ParamValue`s defined in TestModelFixtures was manually
crafted using the constructor of ParamValue, so it didn't contain
a valid list of parameter values as the model is binded actually.

The fixture class is therefore modified to exploit the logic of parameter
parsing, depending on the language and its trait. Moreover, some tests
and fixtures are organized and cleaned up for great readability.