Refactor on ParamValue: Introduce class 'Argument'
Introduced the class 'Argument' to make the argument values of parameters
can be rendered in the templates, depending on the language.

A renderer for Argument is registered, dispatching the actual rendering
on the implemented language traits. For example, long-type integers
are followed by 'LL' suffix in C++ templates, but not in test-cases
templates (issue #151).

Just a note for contributors: To make that our previous templates
work correctly, the type of ParamValue#ValueList is changed to Argument[].